Initial commit of TDM
All of the key parts of the library are here, but many of them are not
fully fleshed out.

The most complete part of the library so far is the OpenAPI "Generator".
This takes in the URL of an OpenAPI 3.0 definition, and generates a set
of fixture definition files in valid Typescript. It is not 100% compliant
with the Open API 3.0 spec, but it covers a lot of it and has been tested
against several APIs including the Stripe API.